Rebase the project architecture for server part and add Passport module #10
