+++ server/package.json
@@ -11,11 +11,16 @@
     },
     "dependencies": {
         "@prisma/client": "^4.8.1",
+        "body-parser": "^1.20.1",
         "dotenv": "^16.0.3",
         "dotenv-cli": "^6.0.0",
         "express": "^4.16.1",
+        "jsonwebtoken": "^9.0.0",
+        "mocha": "^10.2.0",
         "mysql": "^2.18.1",
-        "mocha": "^10.2.0"
+        "passport": "^0.6.0",
+        "passport-jwt": "^4.0.1",
+        "passport-local": "^1.0.0"
     },
     "repository": {
         "type": "git",
